Add files via upload
Arduino controller for a camera Pan and Tilt platform.  For use in a church streaming situation to quickly move to regularly used shots.
Follow mode - to follow joystick pots.  Store mode to save the current position in one of three positions.  Goto mode to drive to one of the three preset positions.
Still needs a proper build and better power supply arrangements to stop servos jittering.